flask/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.2
gunicorn==21.2.0
hurry.filesize==0.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
Werkzeug==2.3.7
